Defer cache invalidation until the very end of every request.
Removes a race condition, that may cause re-rendering of cache items before the database changes causing that re-rendering are visible to the rendering thread.

Requests served by the riot-servlet, are handled by a DeferredInvalidationHandlerInterceptor. That interceptor is put on the list of interceptors via <override:add ... />. Because the order of such adds is undefined when different modules are involved, it's not guaranteed that the DeferredInvalidationHandlerInterceptor is invoked before the OpenSessionInViewInterceptor. For that reason it's important, that all databae changes causing the cache invalidation are flushed before the hibernate session is actually closed. Existing riot code ensures that by modifying the database only inside transactional methods.

For requests served by the website-servlet a filter plugin (DeferredInvalidationHandlerFilterPlugin) is registered. This plugin has highest precedence and will be on the filter chain before the OpenSessionInViewFilterPlugin. Cache invalidation will always take place after any modification is written to the database.

If you need more manual control over the whole process of deferred cache invalidation, use the DeferredInvalidationHandlerFilter and roll your own. All mentioned classes handle nested execution properly. So it's safe to have more than one DeferredInvalidationFilter/FilterPlugin/Interceptor enabled for one single request.