Delayed removal of obsolete live contents when publishing component lists.
When a component list is published, the old live content (if such a content exists) is no longer immediately deleted from the database. Instead it will get removed after a configurable delay. You can use the application property riot.components.delayedContentCleanup.delay to choose an amoutn of seconds, that fit your needs. If omitted a default of 60 seconds ist taken.